@@ -1,6 +1,6 @@
 {
   "name": "front-angle",
-  "version": "0.1.0",
+  "version": "1.1.0",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.7.1",
@@ -12,14 +12,13 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "i18next": "^22.4.9",
-    "react": "^17.0.2",
+    "react": "^19.0.0-rc.1",
     "react-bootstrap": "^2.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.0.0-rc.1",
     "react-i18next": "^12.1.4",
     "react-image-gallery": "^1.2.7",
     "react-intersection-observer": "^9.4.2",
-    "react-router": "5.2.0",
-    "react-router-dom": "5.2.0",
+    "react-router": "^7.12.0",
     "react-router-hash-link": "^2.4.3",
     "react-slick": "^0.28.1",
     "sass": "^1.47.0",
