@@ -1,13 +1,26 @@
-import React from "react";
-import ReactDOM from "react-dom";
-
-import { HashRouter, BrowserRouter} from "react-router-dom";
-
-import Main from "./Main"
-const app = document.getElementById('app');
-
-ReactDOM.render(
-    <HashRouter>
-      <Main/>
-    </HashRouter>,
-  app);
+// import React from "react";
+// import ReactDOM from "react-dom";
+//
+// import { HashRouter, BrowserRouter} from "react-router-dom";
+//
+// import Main from "./Main"
+// import {createRoot} from "react-dom/client";
+// const app = document.getElementById('app');
+//
+// const container = document.getElementById('root');
+// const root = createRoot(container);
+//
+// root.render(
+//     <React.StrictMode>
+//         <HashRouter>
+//             <Main />
+//         </HashRouter>
+//     </React.StrictMode>
+// );
+//
+// // createRoot()
+// // ReactDOM.render(
+// //     <HashRouter>
+// //       <Main/>
+// //     </HashRouter>,
+// //   app);
