@@ -2,11 +2,12 @@ import React from "react";
 
 import HeaderPanel from "./HeaderPanel";
 import RightPanel from "./RightPanel";
-import {Switch, Route} from "react-router-dom";
+import { Route} from "react-router-dom";
 import HomePanel from "./HomePanel";
 import Screenshots from "./Screenshots";
 import CodesamplesPanel from './CodesamplesPanel';
 import "./index.scss"
+import {Routes} from "react-router";
 /**
  * Contains nav on top, footer on bottom, and all the pages are passed in
  * as children..
@@ -24,17 +25,11 @@ export default function PageLayoutICharts(props) {
             <div style={containerStyle}>
                 <div>
                     <div>
-                        <Switch>
-                            <Route exact path="/icharts">
-                                <HomePanel/>
-                            </Route>
-                            <Route path="/icharts/screenshots">
-                                <Screenshots/>
-                            </Route>
-                            <Route path="/icharts/codesamples">
-                                <CodesamplesPanel/>
-                            </Route>
-                        </Switch>
+                        <Routes>
+                            <Route path="/icharts" element={<HomePanel />} />
+                            <Route path="/icharts/screenshots" element={<Screenshots />} />
+                            <Route path="/icharts/codesamples" element={<CodesamplesPanel />} />
+                        </Routes>
                     </div>
                 </div>
             </div>
